Implement lift control in GlobalTeleOp
Lift position is the angle at which the lift will lift the ball. We set it to a pretty high angle so no matter what it will shoot the ball. Lift power is how fast the ball with go into the lift position, which could change the way the ball shoots. Both of these value are subject to change.